Add add and remove test
