+++ src/modules/add.test.js
@@ -2,7 +2,11 @@
  * @jest-environment jsdom
  */
 
+import './__mocks__/mockHtml.js';
+
 import { add } from './add.js';
+
+import displaytasks from './display.js';
 
 describe('test add ', () => {
   it('Add tasks to local storage', () => {
@@ -19,5 +23,16 @@ describe('test add ', () => {
       { description: 'finish', completed: false, index: 4 },
       { description: 'well-done', completed: false, index: 5 },
     ]);
+  });
+
+  it('Add tasks one li to html', () => {
+    displaytasks();
+    const list = document.querySelectorAll('.style');
+    const initiallength = list.length;
+    add('onemore');
+    displaytasks();
+    const finallength = document.querySelectorAll('.style').length;
+
+    expect(finallength).toEqual(initiallength + 1);
   });
 });
