+++ src/modules/remove.test.js
@@ -3,7 +3,12 @@
  */
 
 import { tasks } from './add.js';
+
 import remove from './remove.js';
+
+import './__mocks__/mockHtml.js';
+
+import displaytasks from './display.js';
 
 describe('test remove ', () => {
   it('Remove tasks to local storage', () => {
@@ -18,5 +23,16 @@ describe('test remove ', () => {
       { description: 'finish', completed: true, index: 3 },
       { description: 'well-done', completed: true, index: 4 },
     ]);
+  });
+
+  it('Add tasks one li to html', () => {
+    displaytasks();
+    const list = document.querySelectorAll('.style');
+    const initiallength = list.length;
+    remove(2);
+    displaytasks();
+    const finallength = document.querySelectorAll('.style').length;
+
+    expect(finallength).toEqual(initiallength - 1);
   });
 });
